# Conflicts:
#	lib/runtime-c-api/src/lib.rs

+++ lib/emscripten/build/emtests.rs
@@ -46,26 +46,42 @@ pub fn compile(file: &str, ignores: &Vec<String>) -> Option<String> {
     output_path.set_extension("js");
     let output_str = output_path.to_str().unwrap();
 
-    // Compile to wasm
-    let _wasm_compilation = Command::new("emcc")
-        .arg(file)
-        .arg("-s")
-        .arg("WASM=1")
-        .arg("-o")
-        .arg(output_str)
-        .output()
-        .expect("failed to execute process");
+    let wasm_file_metadata = {
+        let mut wasm_file_path = PathBuf::from(file);
+        wasm_file_path.set_extension("wasm");
+        if let Ok(wasm_file) = File::open(wasm_file_path) {
+            Some(wasm_file.metadata().unwrap())
+        } else {
+            None
+        }
+    };
 
-    // panic!("{:?}", wasm_compilation);
-    // if output.stderr {
-    //     panic!("{}", output.stderr);
-    // }
-    // Remove js file
+    let real_file = File::open(file).unwrap();
+    let file_metadata = real_file.metadata().unwrap();
+    if wasm_file_metadata.is_none()
+        || file_metadata.modified().unwrap() >= wasm_file_metadata.unwrap().modified().unwrap()
+    {
+        // Compile to wasm
+        let _wasm_compilation = Command::new("emcc")
+            .arg(file)
+            .arg("-s")
+            .arg("WASM=1")
+            .arg("-o")
+            .arg(output_str)
+            .output()
+            .expect("failed to execute process");
 
-    if Path::new(output_str).is_file() {
-        fs::remove_file(output_str).unwrap();
-    } else {
-        println!("Output JS not found: {}", output_str);
+        // panic!("{:?}", wasm_compilation);
+        // if output.stderr {
+        //     panic!("{}", output.stderr);
+        // }
+        // Remove js file
+
+        if Path::new(output_str).is_file() {
+            fs::remove_file(output_str).unwrap();
+        } else {
+            println!("Output JS not found: {}", output_str);
+        }
     }
 
     let mut output_path = PathBuf::from(file);
